+++ core/src/com/horse/pokemon/ObjectData/Players/PrimitiveConsumer.java
@@ -2,5 +2,5 @@ package com.horse.pokemon.ObjectData.Players;
 
 @FunctionalInterface
 public interface PrimitiveConsumer {
-    void accept(int f);
+    void accept(float f);
 }
