+++ core/src/com/horse/pokemon/ObjectData/Players/PrimitiveSupplier.java
@@ -2,5 +2,5 @@ package com.horse.pokemon.ObjectData.Players;
 
 @FunctionalInterface
 public interface PrimitiveSupplier {
-    int get();
+    float get();
 }
